Step4 - 요금 조회 (#360)
* refactor(common): 패키지명 변경

* refactor(favorite): 엔티티 필드명 변경

* feat(favorite): 즐겨찾기 삭제 시 소유자 확인

* docs(readme): 요구 사항 추가

* test(path): 지하철 경로 찾기 관련 기능 인수 테스트 수정

* feat(path): 거리별 요금 정책

* feat(path): 노선별 추가 요금 정책

* feat(fare): 연령별 할인 정책

* feat(fare): 요금 조회 서비스 구현

* feat(path): 비회원도 경로 조회 가능

* refactor(path): 변수명 변경

* feat(favorite): station과 연관관계 설정

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,6 +1,7 @@
 package nextstep.subway.favorite.application;
 
 import nextstep.subway.auth.domain.LoginMember;
+import nextstep.subway.common.domain.FindFailedException;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.favorite.dto.FavoriteRequest;
@@ -31,23 +32,25 @@ public class FavoriteService {
     }
 
     public FavoriteResponse createFavorite(LoginMember loginMember, FavoriteRequest favoriteRequest) {
-        Favorite favorite = favoriteRequest.toFavorite();
+        Favorite favorite = new Favorite(
+                findStationById(favoriteRequest.getSource()), findStationById(favoriteRequest.getTarget()));
         favorite.changeMember(findMemberById(loginMember.getId()));
         Favorite savedFavorite = favoriteRepository.save(favorite);
-        return FavoriteResponse.of(savedFavorite,
-                findStationById(savedFavorite.getSource()),
-                findStationById(savedFavorite.getTarget()));
+        return FavoriteResponse.of(savedFavorite, savedFavorite.getSourceStation(), savedFavorite.getTargetStation());
     }
 
     @Transactional(readOnly = true)
     public List<FavoriteResponse> findFavorites(LoginMember loginMember) {
         return favoriteRepository.findAllByMember(findMemberById(loginMember.getId()))
                 .stream()
-                .map(it -> FavoriteResponse.of(it, findStationById(it.getSource()), findStationById(it.getTarget())))
+                .map(it -> FavoriteResponse.of(it, it.getSourceStation(), it.getTargetStation()))
                 .collect(Collectors.toList());
     }
 
-    public void deleteFavorite(Long id) {
+    public void deleteFavorite(LoginMember loginMember, Long id) {
+        Favorite favorite = favoriteRepository.findById(id)
+                .orElseThrow(() -> new FindFailedException(Favorite.class));
+        favorite.validateOwner(findMemberById(loginMember.getId()));
         favoriteRepository.deleteById(id);
     }
